Split GPU ref tests into fast/slow binaries
Extract shared test fixtures into headers and move medium/large
shape instantiations from DISABLED_ tests into a new
hipdnn_gpu_ref_slow_tests binary with integration_test + slow labels.
